backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.12
qrcode==7.4.2
pillow==10.4.0
alembic==1.13.3
psycopg2-binary==2.9.9
